+++ pyads/ads.py
@@ -778,19 +778,18 @@ class Connection(object):
         return_ctypes=False,
         check_length=True,
     ):
-        # type: (int, int, Type, Any, Type, bool, bool) -> Any
+        # type: (int, int, Optional[Type], Any, Optional[Type], bool, bool) -> Any
         """Read and write data synchronous from/to an ADS-device.
 
         :param int index_group: PLC storage area, according to the INDEXGROUP
             constants
         :param int index_offset: PLC storage address
-        :param int plc_read_datatype: type of the data given to the PLC to
-            respond to, according to PLCTYPE constants
+        :param Type plc_read_datatype: type of the data given to the PLC to respond to,
+            according to PLCTYPE constants, or None to not read anything
         :param value: value to write to the storage address of the PLC
-        :param plc_write_datatype: type of the data given to the PLC,
-            according to PLCTYPE constants
-            :rtype: PLCTYPE
-    :param bool return_ctypes: return ctypes instead of python types if True
+        :param Type plc_write_datatype: type of the data given to the PLC, according to
+            PLCTYPE constants, or None to not write anything
+        :param bool return_ctypes: return ctypes instead of python types if True
         (default: False)
         :param bool check_length: check whether the amount of bytes read matches the size
             of the read data type (default: True)
@@ -1057,8 +1056,8 @@ class Connection(object):
         if self._port is not None:
             adsSyncSetTimeoutEx(self._port, ms)
 
-    def notification(self, plc_datatype=None, timestamp_as_filetime=False):
-        # type: (Optional[Type], bool) -> Callable
+    def notification(self, plc_datatype=None):
+        # type: (Optional[Type[Any]]) -> Callable
         """Decorate a callback function.
 
         **Decorator**.
@@ -1069,11 +1068,6 @@ class Connection(object):
 
         :param plc_datatype: The PLC datatype that needs to be converted. This can
         be any basic PLC datatype or a `ctypes.Structure`.
-        :param timestamp_as_filetim: Whether the notification timestamp should be returned
-        as `datetime.datetime` (False) or Windows `FILETIME` as originally transmitted
-        via ADS (True). Be aware that the precision of `datetime.datetime` is limited to
-        microseconds, while FILETIME allows for 100 ns. This may be relevant when using
-        task cycle times such as 62.5 µs. Default: False.
 
         The callback functions need to be of the following type:
 
@@ -1107,7 +1101,7 @@ class Connection(object):
         """
 
         def notification_decorator(func):
-            # type: (Union[Callable[[int, str, datetime, Any], None], Callable[[int, str, int, Any], None]]) -> Callable[[Any, str], None] # noqa: E501
+            # type: (Callable[[int, str, datetime, Any], None]) -> Callable[[Any, str], None] # noqa: E501
 
             def func_wrapper(notification, data_name):
                 # type: (Any, str) -> None
@@ -1142,12 +1136,9 @@ class Connection(object):
                         0
                     ]
 
-                if timestamp_as_filetime:
-                    timestamp = contents.nTimeStamp
-                else:
-                    timestamp = filetime_to_dt(contents.nTimeStamp)
+                dt = filetime_to_dt(contents.nTimeStamp)
 
-                return func(contents.hNotification, data_name, timestamp, value)
+                return func(contents.hNotification, data_name, dt, value)
 
             return func_wrapper